Fixed compilation errors on Linux Enabled whole program optimization
Assets/BSP:
- Moved code related to grouping surfaces in a new class
- Removed unnecessary fields from some structures

Assets/Image:
- Now throw the correct exception instead of the generic ImageException exception

Assets/SkeletonAnimation:
- Fixed the import of "newanim" animations
- Removed the useless **Skeletor** class

Assets/TIKI:
- Added support for conditional includes (for example it is possible to only allow includes with 'utils')
- All parsing errors seem to be fixed now
- Now preallocate one block of memory for the entire TIKIAnim data

Files:
- Removed FileManager class in favor of an IFileManager interface to split pak-level files from OS-level files
- Made PakFileManager (for handling pak files) and SystemFileManager classes (for handling OS files)
- Improved support for categories (they are now hierarchical)

Utility:
- Removed Script in favor of TokenParser

+++ include/MOHPC/Network/Remote/Socket.h
@@ -31,7 +31,7 @@ namespace MOHPC
 			 * @param	timeout		Time to wait. (0 = no wait)
 			 * @return	True if there is any pending data.
 			 */
-			virtual bool wait(size_t timeout) = 0;
+			virtual bool wait(uint64_t timeout) = 0;
 
 			/**
 			 * Return the number of pending bytes waiting in socket.
